Fix --derive-all in drawdag
Summary: The `--derive-all` argument that's used with `drawdag` tool in integration tests doesn't do what it's supposed to do. It doesn't derive all the known derived data but instead ends up deriving a bunch of hard coded types. Thanks to pierrec's `bulk_derivation` library, we can now get rid of the hardcoding and ensure that we always derive all known derived data types that are enabled for the given repo.

Differential Revision: D58865122

fbshipit-source-id: b17fed83b0d22683c6d4423d0ad8ee0374de86f5