Live Git Sync: Improve the support for BOM bytes in commit messages
Summary:
When decoding a string that contains a BOM, the previous implementation (using `encoding_rs::decode`) would loose the Byte Order Mark.

That means that we get a lossy conversion on valid UTF-8 inputs, which can be avoided.

Instead, use the BOM to determine the encoding if present (which `decode` already did), but don't remove the `BOM` from the output.

This is useful as some tools on Windows like to add a BOM mark to UTF-8 encoded strings, however futile the endeavour.

Reviewed By: RajivTS

Differential Revision: D58871944

fbshipit-source-id: 3fc1d23726c926e47bb8bc115e64ee07d0e52c8e