Live Git Sync: Refactor decoding tests
Summary:
The way the tests were written was a bit too abstract and hard to read as one
would need to keep jumping between the const block and the tests to understand.

Instead, explicitely write the strings in-place and add comments where helpful to
help with readability.

I'm doing this change because I'm introducing more tests in the next diff and I want them
to be both readable and consistent with existing tests.

Reviewed By: RajivTS

Differential Revision: D58871945

fbshipit-source-id: 218264b7853eb04d2542bb5c46047073a943d6bd