Replace '/' with '.' in TARGETS files
Summary:
There are many Python targets using a file path (contains '/' or '.py' file extension) instead of canonical module name (uses '.' instead of '/') as the `module_name`. It brings complexity to existing Buck macros, and also risks to update or write new macros.

This change unifies the usage to module name instead of file path in 3 Diffs:
1. replace `/` with `.` in TARGETS `main_module`
2. strip ending `.py` in `main_module`
3. update Buck macros

This is the 1st Diff.

The changes are generated by N5492699.

Reviewed By: zsol

Differential Revision: D58565964

fbshipit-source-id: f1e93a5836f21a7420c3f513b20345ddfa30664e