enable edenapi push for test-pushrebase-logging.t
Summary:
The new push has two commit entries in scribe log:
* one is created by upload changesets api
* another is created by land_stack api

As Mark commented below: "In wireproto-based push, we did "upload changeset; pushrebase; log" all as one action, so we were able to skip the logging of the uploaded draft.

Now, uploading the changeset and pushrebasing are separate operations, so we log the draft as well as the pushrebased public commit.

The fact that we skipped the draft before was a bit of a dubious opimization anyway, so I don't think there is harm in including it now."

Reviewed By: markbt, andreacampi

Differential Revision: D58791691

fbshipit-source-id: 780ef19a3013431c516377aa2d1cc9586dbcd09d