enable edenapi push for test-server-post-push-logging.t
Summary:
The new push has two commit entries in scribe log:
* one is created by upload changesets api
* another is created by land_stack api

As markbt commented in D58791691: "In wireproto-based push, we did "upload changeset; pushrebase; log" all as one action, so we were able to skip the logging of the uploaded draft.

Now, uploading the changeset and pushrebasing are separate operations, so we log the draft as well as the pushrebased public commit.

The fact that we skipped the draft before was a bit of a dubious opimization anyway, so I don't think there is harm in including it now."

Reviewed By: andreacampi

Differential Revision: D58817183

fbshipit-source-id: 279fa8488d1012d5b713d5c912f18eefc0648047