sharded_map_v2: make node weight limit part of `ShardedMapV2Value` trait
Summary: Currently the weight limit of sharded map nodes is specified in the thrift file. This means that the weight limit for all sharded map types is the same. This diff moves it to be part of the `ShardedMapV2Value` trait instead to allow specifying different weight limits and to force new type implementers to consider what the appropriate weight limit for their type should be.

Reviewed By: RajivTS

Differential Revision: D57061282

fbshipit-source-id: 8eafa4449ed3a3608920e510e7cda77dddbdf5e0