Use the V2 logger config for updated schema and incremental mode
Summary:
We want the rust `RepoMetadataLogger` to run in incremental mode. However, this is a new tool that is meant to replace the existing python script so we won't be making a direct switch and instead will have a period of side-by-side execution.

D58946992 introduces the new logger config and this diff makes use of that logger config.

Reviewed By: YousefSalama

Differential Revision: D58946573

fbshipit-source-id: b622a5d883df9ea2ab7fff5509a1e0a3b57b2def