Fix bug from deleting submodules with recursive submodules
Summary:
Fixes the bug described in D58954457.

## Current implementation
The current implementation queried the submodule repo to get all the files (i.e. leaves) present in it at the pointed commit.
The problem is that doing the same for the recursive repos would be **way more complicated**, because we would need to get their pointers in the first submodule and so on. This would introduce a lot of recursive calls and make the code a lot more complicated (I tried that initially).

## New solution
The new solution is simpler, way more elegant and actually fixes the bug.

The goal is to generate all the paths (need to be files) that need to be deleted. To do that, I can just re-use the function that expands submodules (`expand_git_submodule_file_change`), by calling it with the submodule pointer from the parents and passing an empty set of parents to it. This generates the entire working copy of the expansion, because the `submodule_diff` function, that generates the delta, will have no previous commit to compare to.

Basically, if we have commits `A - B - C - D`
- B introduces submodule
- C modifies submodule
- D deletes submodule.

I call `expand_git_submodule_file_change` passing the submodule pointer of C and setting its parents to an empty set.
This will generate the submodule expansion working copy in commit C.

Reviewed By: RajivTS

Differential Revision: D58953176

fbshipit-source-id: 7b819a913a743d34fb150a986678eaaae3005ad5