Back out "Back out "[mononoke] commit_graph: add new prefetching target for fetching exact skip tree ancestors""
Summary:
We did the first backout for mitigating S428805, but on further inspection, the recursive CTEs that overloaded MySQL were different from these.
Also this is behind a jk, so we have a less heavy handed way to deal with this if it causes problems.

Original commit changeset: 90b00b447545

Original Phabricator Diff: D58991192

Reviewed By: YousefSalama

Differential Revision: D58996765

fbshipit-source-id: c581ccd401a79217b810494828df78f69923c977