show split banner on amends
Summary:
this diff adds the split suggestion banner to amends.  it also takes into account when an amend would have a net effect that would reduce SLOC and it won't show the suggestion in that case.

we do however keep showing the SLOC as-is to make things easy to reason about.

Reviewed By: evangrayk

Differential Revision: D58562791

fbshipit-source-id: b4724e18d9da60e5a684980e6fdd81952126c1c9