define optimisticRevset in CommitInfo to resolve optimistic commits
Summary:
This is a redo of D58849893.

ISL is based on a queueing system, so you can run commands one after another and act on the optimistic preview of the commit DAG. This works great in most cases by using revsets that automatically look for the latest successor.

However, certain commands create new commits entirely, such as sl commit. Such commands must produce fake commit hashes for their optimistic commits. If you try to act on one of these fake optimistic commits, it will dutifully pass the optimistic commit's hash to whatever sl command, which will always be an error.

This diff attempts to correct this fundamental issue by introducing a notion of resolving fake optimistic commit hashes.

The idea is:
- have the optimisticDag in the Operation add a field to CommitInfo which includes a revset that will match exactly the newly created optimistic commit made by this operation
- use that revset instead of the commit hash when convering a commit to be run as a later operation's args

This approach is admittedly slightly fragile. Writing the revset to match the result of the operation is a little difficult. For the commit operation, we need to take the earliest descendant of the parent commit created after the commit operation was started. Writing this for EditStack might be similarly tricky.

Eventually, I think we should move to having a communication channel with operations that call tell us exactly the replacement hash for an optimistic commit. This will let us simply do the replacement on the server.

Reviewed By: quark-zju

Differential Revision: D58980031

fbshipit-source-id: e9fa57f6764625e206ede11e53b6e10589d6de40