Introduce OptimisticRevset
Summary:
In the previous diff, it became possible to know a revset that describes the replacement for an optimistic commit.

In this diff, we formalize this idea so that there's a new variant of `SucceedableRevset` / `ExactRevset`, called `OptimisticRevset`. This describes a revset for a commit that does not exist locally, because it will be created by an operation. This CommandArg is used for cases when you try to use such a fake commit as an input for a queued command.

Adding this third value requires most callers to learn how to handle this case. It's slightly different from the others since `.revset` now refers to a value that is most certainly not understandable as a hash in the UI, and `.fake` is how you get to the value that represents something like a Hash.

This distinction matters since we cache a lot of information by Hash. For example, we need the optimistic preview for the Hide command to set the preview type for the fake commit hash, so that we see the "run hide" button appear.

This diff is mostly just going through Operations and supporting the OptimisticRevset.

There's a couple of cases where we don't conveniently have the CommitInfo available to make or understand an OptimisticRevset. We'll need to fix these. Especially JfSubmit.

Note: I considered making OptimisticRevset's fake hash be in the "revset" field, so most users can use `.revset` regardless of if it's OptimisticRevset or SucceedableRevset or ExactRevset. But I think I prefer the clarity in the name with "revset" and "fake" (or something similar)

Reviewed By: quark-zju

Differential Revision: D58980032

fbshipit-source-id: 692bd08caaa0fd6772db59695004a91ad0433290