ephemeral_blobstore: implement ephemeral commit graph storage
Summary:
The `Changesets` trait is implemented for `EphemeralChangesets` by relying on three things:
- The bonsai changesets stored in the ephemeral blobstore. This is used to find the parents of a changeset in a snapshot.
- `ephemeral_bubble_changeset_mapping` SQL table. This is used to find what changesets are inside of a bubble as well as their generation numbers.
- The persistent `changesets/csparents` SQL tables for information about changesets that are not in bubbles.

In order to extend the commit graph to support operations involving snapshots, we need a commit graph storage with a similar structure. The main challenge is that the commit graph calculates many attributes for each added changeset (e.g. skew binary edges) that involves traversing parts of the commit graph, and we don't want to duplicate this logic.

To avoid duplication I'm doing the following for EphemeralCommitGraphStorage:
- The storage contains a MemWrites wrapper around the persistent commit graph storage. All inserts write to the `ephemeral_bubble_changeset_mapping` SQL table first, then to the MemWrites memory storage (writing to the MemWrite storage here is an optimization).
- All reads go first to the MemWrites storage. If it's not found there we use the `CommitGraph::add_recursive` method to add the missing changeset and all of its missing ancestors to the MemWrites storage using the ephemeral blobstore to traverse the part of the graph inside the bubble.

There are other solutions but this seems to be the easiest in terms of not having to create and backfill new SQL tables, and not having to change logic if we ever want to support bubble with more than one changeset.

Differential Revision: D58954276

fbshipit-source-id: d734ceb2c22fb48ebbc382fd9a079f796399f1dd