implement log-only mode for hooks
Summary: To allow safe roll-out of hooks that could potentially impact traffic, add "log-only" mode to hook config.  When set, the hook is still run as normal, but hook failure does not prevent landing.  The failure message is logged to the scuba table for hooks (in a new column) so that it can be examined, but hook processing proceeds as if the hook passed.

Reviewed By: gustavoavena

Differential Revision: D59004553

fbshipit-source-id: 87ba9d30800a677eb696a53361cd862ff22479bd