git/protocol: use dynamic dispatch in `GitDeltaManifestOps`
Summary: Static dispatch doesn't work well for `GitDeltaManifestOps` trait because we want to decide at runtime whether to use `GitDeltaManifest` or `GitDeltaManifestV2`. We could probably refactor the code so that we have a match statement at the beginning of each protocol method, but I don't think it's worth it. This diff modifies the trait and the protocol code to use dynamic dispatch instead.

Reviewed By: RajivTS

Differential Revision: D59010606

fbshipit-source-id: e5198dc2d3748da2acf85fa095b33c9d3c52b1bc