commit cloud: Fix versions conflict handling
Summary:
I was a adding a full commit cloud test and noticed a failure after updating a workspace version. Basically what happened was  that `CURRENT_TIMESTAMP` in sqlite returns a string-like date e.g `2024-06-..` while we stored timestamps as bigints (unix timestamps).

For fixing I'm simply replacing `CURRENT_TIMESTAMP` with `Timestamp::now()` and passing that to the sqlite query. Since we needed to use it in the mysql query aswell for `monononoke_queries!` to compile I simply embedded it into a `COALESCE` operation so if we send an empty timestamp it will take the current one instead, this is unlikely to happen but I figured it was the best use for the variable, suggestions welcomed.

Reviewed By: gustavoavena

Differential Revision: D59064926

fbshipit-source-id: e61ca589584325e37bc4ad06836e84d1950b0652