implement support for eagerepo: part #4 (scmstore)
Summary:
implement support for eagerepo: part #4 (scmstore)

this part contains support for scm store

I've discover that in sapling it is strict expectation that the key from the request and response exactly match for "trees", so that if the path is filled in former, should be in latter. So, we have to adjust a converted tree to include the correct path in its key.

Reviewed By: quark-zju

Differential Revision: D59027701

fbshipit-source-id: b26a70e00c411aba08263835fbc6df4806e60f85