Implement advertisement support for Git write path
Summary:
This is the first diff that begins implementation of Git write protocol. The very first change needs to be made in the advertisement endpoint where the read and the write path vary quite a bit. The integration test shows that we are able to advertise our capabilities and refs and the push only fails when it actually tries to send the packfile containing the push contents.

In follow up diffs, I will add the `receive-pack` endpoint to accept and apply pushes on the server.

Reviewed By: andreacampi

Differential Revision: D59228164

fbshipit-source-id: 84805164358d0467a553e096003d3461571f2e54