scs_server: add boilerplate for commit_linear_history method
Summary:
Context: For querying commit history from SCS we have the commit_history method. However, depending on the parameters passed to this method, it can be very inefficient. Most of the inefficiency comes from having to handle merge commits in the history, but most of our clients do not actually care about merges. Therefore I'm going to introduce another method that ignores merge commits, which allows it to be implemented in a way that's efficient for all possible parameters (efficient as in O((output size) * log))

Adds a new method for querying the linear ancestors of a commit. Linear ancestors of a commit are commits that can be reached by repeatedly following the first parent, including the commit itself. I'm explicilty ignoring the `after_timestamp` and `before_timestamp` parameters because they are well known to cause users to write inefficient queries.

Reviewed By: markbt

Differential Revision: D59273639

fbshipit-source-id: bd5c92b7b46b56637e91600f0a3c73aa084f4a58