cross-repo sync: Break the stack size
Summary:
The cross-repo sync libraries construct really large futures that are on the border of causing stack overflows.

With D58202861, some shuffling of how derived data gets evaluated brings it over the limit and causes stack overflows in unit tests and integration tests, which manifests in SEGFAULTS.

This diff breaks down the stack in all the API functions that the cross-repo sync exposes (forward and backward sync), by boxing large futures.

Reviewed By: YousefSalama

Differential Revision: D59329111

fbshipit-source-id: 73037da9a06b69d47d1fe6ffbaffc4d1e7e5c775