commit_graph: implement a stream for linear ancestors
Summary:
Implements a builder for streams of linear ancestors. The builder allows constraining the stream to exclude the linear ancestors of a changeset, or to include only the linear descendants of a changeset. It additionally allows efficiently skipping the first N linear ancestors of the stream.

This will be used to power commit_linear_history SCS method.

Reviewed By: markbt

Differential Revision: D59324016

fbshipit-source-id: 8d5f3ac9f3f13d516457449cf28c916b02e20a8d