config: Add configurable derivation batch sizes
Summary:
In the child diff, we replace the implementation of `derive` with a new version that sections the work into batches and derives in batches.

It will be useful to configure the behaviour so that for each individual derived data type, we can specify a reasonable batch size.

Add the config mechanics now and use them in the child diff.

Reviewed By: YousefSalama

Differential Revision: D59329112

fbshipit-source-id: fc58e2671163e8657b295f684135bf100c980955