derived_data: Always derive efficiently
Summary:
Before this change, there were two ways to interact with data derivation:
* `derive_exactly_batch`: a function that can be used efficiently but requires a lot of care (for instance, knowing for sure that the parents of each commit in the batch are already derived and ensuring the ordering in the batch is topological)
* `derive`: a function that's easy to use, and "does the right thing", but does so very inefficiently, with no batching and "O(n_commits * derived_data_size)" memory usage

The rule of thumb was: practically never call `derive` except if you're sure it's delegating to the derived data service that does the right thing of calling the batched interface or you will cause a SEV eventually.

Reconcile both of these implementations and sprinkle some `CommitGraph` magic on top so we can have the best of both worlds.

NOTE: We sacrifice a tiny surface of functionality, namely, logging the time spent finding underived commits, but we also make it so much more efficient that it really shouldn't matter anymore)

Reviewed By: YousefSalama

Differential Revision: D58202861

fbshipit-source-id: f30406ef9a8f0042e0f30ff9ea6f73184de775f8