Implement packfile parsing capability in Mononoke GRit server
Summary:
A key part of the `git push` payload is the `packfile` which contains the all the Git objects that the client wants to send over to the server. This packfile is encoded, compressed and deltified.

This diff adds the capability to parse the packfile and verify its validity. It produces a stream of objects that need to be persisted in Mononoke blobstore (which will be handled in a follow-up diff)

Reviewed By: andreacampi

Differential Revision: D59332462

fbshipit-source-id: 5c7b24649c17293c4bb69194dcfae295c9de621d