add integration test
Summary:
# Context

We'd like to add log rotation for EdenFS (and eventually Watchman, Myles, etc). There are services that already exists on Linux and macOS (logrotate and newsyslog respectively). However, these services require our daemon's to implement log-reopening logic in order to work correctly (on Linux) or at all (on macOS).

We could deal with slightly correctness issues on Linux, but since we need log-reopening logic for log rotation to work at all on macOS, we should implement log-reopening so that we can use log rotation on at least two of our environments.

# Implementation

The basic way to implement log-reopening in Eden is to add a signal handler for SIGHUP. SIGHUP is used to indicate "hanging up" (in other words, the file we're writing to is now closed). When we receive SIGHUP, we can reopen the log file that we're writing to so that we can continue to persist our logs.

# This diff

This diff adds an integration test to show that SIGHUP handling doesn't work right now. The next diff should fix the test and prove that we handle SIGHUP correctly (and it fixes writing to a closed log file).

Reviewed By: chadaustin

Differential Revision: D59303688

fbshipit-source-id: dbbd83ba7078995918fd592d78bcae95f574e183