add stub for SIGHUP handler
Summary:
# Context

We'd like to add log rotation for EdenFS (and eventually Watchman, Myles, etc). There are services that already exists on Linux and macOS (logrotate and newsyslog respectively). However, these services require our daemon's to implement log-reopening logic in order to work correctly (on Linux) or at all (on macOS).

We could deal with slightly correctness issues on Linux, but since we need log-reopening logic for log rotation to work at all on macOS, we should implement log-reopening so that we can use log rotation on at least two of our environments.

# Implementation

The basic way to implement log-reopening in Eden is to add a signal handler for SIGHUP. SIGHUP is used to indicate "hanging up" (in other words, the file we're writing to is now closed). When we receive SIGHUP, we can reopen the log file that we're writing to so that we can continue to persist our logs.

# This diff

This diff adds a stub for SIGHUP handling so that I can write an integration test in the next diff that avoids crashing the Eden daemon (sending SIGHUP to a process that doesn't handle SIGHUP will cause that process to exit).

Reviewed By: chadaustin

Differential Revision: D59303687

fbshipit-source-id: dd924957beac71868aea3adea37992733de64df3