Optionally hook up the DB to the read path
Summary:
Rather than trying to make the switch in one go, I'm introducing a new constructor. This will allow me to add some callsite at a time and test thoroughly. As long as the content of the DB and Configerator match, this should not cause any observable behavior, regardless of which code path is taken.

When the DB is empty (which is the case now) this is a giant noop anyway.

Differential Revision: D59367428

fbshipit-source-id: 08e9ac0734ad280381379ce8089bc214489e808c