manifest: implement SortedVectorTrieMap
Summary:
Implement a wrapper around `SortedVectorMap` that simulates a `TrieMap` without actually constructing one.

When using unsharded manifests in the sharded manifest algorithms, they need to provide the `TrieMapOps` interface.  Currently we do this by converting the unsharded manifest to a fully-expanded `TrieMap`.  However, for large manifests, this is expensive, and often wasteful as large parts of the manifest are likely the same anyway.

As an alternative, we can use the original `SortedVectorMap` that the unsharded manifest is built on as a basis for the trie map traversal.  We track the length of the common prefix so far traversed, and the start and end positions of all entries within the sorted vector map that start with that common prefix.  The individual trie map node emulators can share the original sorted vector map, so no copying of data is required (except to copy out values when they are reached).

Note that `SortedVectorTrieMap` is read-only.   It can be thought of as a trie-map "view" onto the original map.

Reviewed By: andreacampi

Differential Revision: D59380491

fbshipit-source-id: be2b52aad2673361b7151ad6492de0dcdaaa12d9