add TreeMetadata structures
Summary:
# Context

The new AugmentedManifest format used by Sapling and Mononoke will support looking up Trees/Blobs from CAS via (Blake3,size) pairs. We already provide a way for users to get (Blake3,size) pairs for blobs, but there is currently no easy way to get this information for trees.

We want to add support to getAttributesFromFilesV2 for looking up the (Blake3,size) pairs for trees as well. To do this, we'll need to add a method for querying TreeMetadata (including size and Blake3 hashes) via the SaplingBackingStore.

This initial implementation of TreeMetadata will only include the following features:

1) Ability to query TreeMetadata from the SaplingBackingStore via a getTreeMetadata endpoint
2) The ability to request TreeMetadata from the getAttributesFromFilesV2 thrift endpoint

In the future, we will extend the implementation to support:

1) Automatically fetching TreeMetadata during BackingStore::getTree requests
2) Caching TreeMetadata in Eden's in-memory caches

# This diff

This diff adds the strucutres that we'll use to pass around TreeMetadata information. This currently only includes the directory size (in bytes) and optional Blake3 hash, but it could be extended in the future.

Reviewed By: jdelliot

Differential Revision: D58418093

fbshipit-source-id: c1de7220053be90cc0c8f627e9c6a924a50e8263