derived_data_manager: add bubble id to DeriveRequest
Summary:
Adds the bubble id of the manager to `DeriveRequest` which is sent to the derived data service when calling `derive_remotely`. The bubble id is required for the derived data service to be able to see the snapshot during derivation.

This is ignored on the derived data service side for now.

Differential Revision: D59369487

fbshipit-source-id: 0544f953ca22a41054f47f00b437ab9ac8510493