Integration tests support for XDB push redirection
Summary:
To properly test the multiple code paths in D59367428 we'll needs JKs. `setup_common_config` is a good place to initialize them, except not all tests call it.

This change is a noop, because the JKs are set to the default value. The next diffs in the stack will enable the XDB mode and update tests to match.

Reviewed By: RajivTS

Differential Revision: D59391188

fbshipit-source-id: 49240ec93a87af91fd6ce122406ce32307ec7e22