Run tsc on isl folders in CI
Summary:
We seemingly weren't doing typechecking in our CI. I think this used to happen incidentally from `jest`, but we turned that off for local runs to work better. It doesn't seem like anything bad snuck through, but it's better to do typechecking on CI.

The one type fix on this diff is to fix an error that only appears in OSS during typecheck, since the Internal object essentially becomes `any`, we need to add an annotation.

Reviewed By: quark-zju

Differential Revision: D59344896

fbshipit-source-id: 5be9964c63a1eb16f6f8e1ba0459e32a751ce7e6