config: add check-in snapshot of remote dynamic config
Summary: Add a fallback logic to a snapshot of remote dynamic config if it fails to read it from remote and cache path. Currently, the check-in snapshot is a const string and we need to manually update it, I don't like this solution, but this is fast way to unblock users. In the following diffs, I will add mechanisms to automatically (or semi-automatically, such execute a command) update it.

Reviewed By: muirdm

Differential Revision: D57312334

fbshipit-source-id: a36349ad8c0226675015efad8067a8b3d830c243