Introduce the GitReader trait in GitImport and RemoteGitImport
Summary:
Currently, both `gitimport` and `remote-gitimport` depend upon on the `GitRepoReader` object which works on top of a checked out repo on disk. We want to leverage the same functionality in supporting git pushes but our Mononoke servers are non-checkout based.

This diff introduces the `GitReader` trait and updates all call sites to depend on this trait instead of any concrete implementation. It also restructures the code a bit so all method related to fetching and parsing objects from the git store are supported by this trait. In Mononoke server, this trait will be implemented by a type that will house the objects in memory.

Reviewed By: andreacampi

Differential Revision: D59469723

fbshipit-source-id: f0f07b266c7a926dafdb12819fa79e2fb2bac52f