Refactor gitimport objects library
Summary: The importing of commits, trees, blobs and tags that happen as part of `gitimport` also needs to happen in case of `git push`. To ensure that we can reuse that functionality, let's factor out the `import_commit_contents` function that can be used independently. In a follow up diff, I will remove the hard dependency on `GitRepoReader` and replace that with a trait that can work in the absence of an ondisk repo.

Reviewed By: mitrandir77

Differential Revision: D59467355

fbshipit-source-id: 6377a45999429b80a614eb843cdc1366b091e17f