derived_data_manager: add bubble_id to the primary manager and override primary changesets and commit graph
Summary:
Overriding the changesets and commit graph facets for the primary manager allows it to see ephemeral changesets in the bubble. This is harmless since the manager doesn't write to either facet and it simplifies the logic required to allow the derived data service to derive ephemeral changesets.

Adding the bubble id to the primary manager is also harmless as it's only used when sending derivation requests to the derived data manager to inform it that the derivation is happening in the context of a bubble.

It's only the blobstores really that need to be different between the primary and the secondary managers.

Reviewed By: andreacampi

Differential Revision: D59470516

fbshipit-source-id: 398f1dff62d36cb4445bf764a3c8f5e357643766