goto: fix dirstate out-of-sync issue calling Rust goto
Summary:
If the repo.dirstate._map hasn't been accessed/loaded, then calling into the Rust goto library was leading to treestate<->dirstate out-of-syncness. The in-memory treestate changes from Rust propagated to Python, but the dirstate p1/p2 update don't propagate, so the in-memory treestate and on-disk dirstate became out of sync for Python. Fix by accessing the dirstate._map explicitly.

This is a hacky fix. I couldn't come up with a better, easy fix. Maybe we could drop the p1/p2 from the "dirstate" file, but various things (like "sl whereami") still depend on it.

Reviewed By: quark-zju

Differential Revision: D59400891

fbshipit-source-id: a24b9792cbc35ee14aa074f6f64c35b9b4720728