scmstore: always use Rust repo's file scmstore
Summary:
There was one place in Python we were instantiating a filescmstore directly instead of using the repo's. It was a place that wanted a "shared only" store (i.e. writes only go to cache, not the local/repo store). Tweak the filescmstore::getsharedmutable() method to return a filescmstore instead of a mutabledelta store. I think the former implements all the same methods as the latter (to satisfy the existing caller), and can also serve as a "shared only" filescmstore deisred by makesharedonlyruststore().

This was the last caller of make_filescmstore, so I can kill that and we are down to a single place that instantiates a filescmstore.

Reviewed By: liubov-dmitrieva

Differential Revision: D55004347

fbshipit-source-id: 181d61fa8d9e8467273bec60be8738673d5c3392