scmstore: implement history store traits for TreeStore
Summary:
Implement the read and write history store traits for TreeStore. The read implementations make sense since the TreeStore is where we fetch trees from edenapi (and edenapi provides the parents). The write implementation isn't required, but it allows the TreeStore to be a one-stop-shop for reading and writing tree related stuff.

Nothing uses this yet.

Reviewed By: quark-zju

Differential Revision: D55004352

fbshipit-source-id: 952523ba92087d0caf3809fdc87c2fdcb5bc1218