scmstore: prepare for tree history/parents support
Summary:
I've realized that manifest parent information is currently only fetched over the wire protocol. It is available over SaplingRemoteAPI, but our scmstore TreeStore ignores the data.

The first step to supporting parent info from SaplingRemoteAPI is add knowledge of the "parents" data type to TreeAttributes. This will allow users of the TreeStore to request tree content and/or tree parent info.

Reviewed By: liubov-dmitrieva

Differential Revision: D55004355

fbshipit-source-id: 66cb3b2f7009045b438af1310649aa885fc5b86d