scmstore: hook up tree parent fetching
Summary:
Now in TreeStore::fetch_batch, if parent info was requested we will check the local and shared history stores. If not found, we will then hit edenapi and write the parent info back into the history cache.

Nothing requests parent info yet.

I'm not sure if it makes sense to "prefetch" and cache the tree parents if they weren't requested. I'm thinking not because I don't think tree parents are used that much. I added a config to turn it on if it turns out to be useful.

Reviewed By: liubov-dmitrieva

Differential Revision: D55004354

fbshipit-source-id: d3ccd1b9aff57b9851418bd39efd2bc355ccf73c