derived_data_service: use config to disallow deriving some types for ephemeral bubbles
Summary:
Related configerator diff: D59515024

Some derived data types are not allowed to be derived for ephemeral bubbles. At the moment these are the hg and git derived data types. If a service request a derivation for such type the derived data service will enqueue it in the derivation queue and it will become stuck there. This diff denies deriving such types to avoid this issue.

Reviewed By: RajivTS

Differential Revision: D59515702

fbshipit-source-id: 5d7c5520647c79db1f50190bd9003fe94b76fb96