dag: add `managed_virtual_group` state
Summary:
The VIRTUAL group is special as it does not persist to disk - it gets cleared
after reloading from disk. To make it easier to use, add a state to explicitly
define the desired content of the VIRTUAL group.

This diff just adds the state. The next diff makes it useful.

Reviewed By: muirdm

Differential Revision: D59300592

fbshipit-source-id: 2212e964145c6c6d9db5c6e3ec87899232ee7066