Improve doctor logging to collect stats on number of failed and manual fixes
Summary:
# Context
Eden doctor is a CLI utility that attempts to identify and fix issues that may be causing eden to not work properly.

# This Diff
This diff adds a higher level of detail to the logs collected by eden doctor. It adds 2 categories of logs.
 - Number and type of fixes that were attempted and failed to be fixed automatically
 - Number and type of fixes that require manual intervention by the user.

In addition eden doctor now tells the user when it detects an error it doesn't know how to fix.

Reviewed By: MichaelCuevas

Differential Revision: D59470633

fbshipit-source-id: f52dfc0f7f9a0f347c44b880960a969c669d716f