commit cloud: send notification on workspace update
Summary:
This enables automatic syncs for commit cloud. So whenever something changes in your workspace, a notification is sent so other local copies of that workspace  automatically pulls the changes.

A condition was added so this doesn't run in sandcastle for .t tests. Unsure if we want this for CI at all.

Reviewed By: liubov-dmitrieva

Differential Revision: D59301460

fbshipit-source-id: 0a20a3dd12caffef858782a4d77460c57a3112e1