derivation_worker: spawn coordinators for all derived data configs
Summary: Right now we only spawn a single coordinator for derived data config with name "default". This means that any derivation request that uses a different config would not get processed. This diff changes the logic to spawn a coordinator for each derived data config instead. This should be cheap to do after D56939625, because if the queue for a config is empty the coordinator will set a zelos watch and sleep until a new item is enqueued, so most of the time a single coordinator will be active except for migration periods.

Reviewed By: RajivTS

Differential Revision: D59577917

fbshipit-source-id: ca86fdb1e526253d6650befb00afd5bcf9cad045