Add explicit check for unheatlhy edenFS
Summary:
If we get an error from `sl root`, we assume it's not a repo and suggest cloning. In the case that you're using EdenFS and it's unhealthy, it would appear in the same way, but of course the remediation is different: we should actually tell them to run `eden doctor` (presumably to restart eden, but potentially to fix other issues too)

To do this, we just look in the cwd for README_EDEN.txt. I think this is a decent check, though perhaps we should instead use `eden status` or similar. I could imagine if cwd is a subdir of the repo, this wouldn't fire properly...

Reviewed By: alichtman

Differential Revision: D59544447

fbshipit-source-id: 9640c94a15dd458e527fa9af4213ca6b0ad8d73c