goto: wrap call to Rust goto in dirstate.parentchange()
Summary: dirstate.parentchange() prevents the dirstate from getting written out if there are exceptions during the body. Not 100% sure it is required, but it seems prudent to use around the Rust checkout since there is some division of work (e.g. Rust updates the treestate parents, but the Python updates the dirstate parents, which we don't want to get out of sync).

Reviewed By: quark-zju

Differential Revision: D59400888

fbshipit-source-id: 044c1a736374ae7fe2b6553b643e63ec76b92198