goto: always call treestate.invalidate_mtime
Summary:
Tweak the Rust checkout to always invalidate treestate entries with mtime >= the max mtime of update files. This ensures on-disk changes to the files in the same second will be picked up by future "status" calls.

Previously invalidate_mtime was only called in Rust when we did the full dirstate flush, but to interop with Python better the Rust checkout does not perform the dirstate flush anymore (when called from Python).

Reviewed By: quark-zju

Differential Revision: D59400890

fbshipit-source-id: 3faf4952d29eea62977bd47aaf6e6a8ffb7b6f95