goto: fix "--clean" bug with removed/deleted files
Summary:
If you had a locally removed/deleted file, "sl go -C DEST" would not remove the file from the treestate if it didn't exist in DEST. The treestate updates are based on the diff between the working copy and DEST, and there is no diff in this case.

Add an extra check to notice removed/deleted files that weren't updated during "sl go -C" and remove them from the treestate.

Reviewed By: sggutier

Differential Revision: D59400885

fbshipit-source-id: 1c09a84d829873d07b60bb2c834cfa1cec22ee59