Copy components to their own package dir
Summary:
This starts a stack of commits migrating from components inside isl/src to living in a new top-level components/ dir. This makes it easier for others to share code from our codebase. This new components/ dir will be largely separated from our codebase, and only contains reusable components that aren't isl-specific.

This will be like our own component library for ISL.

This first diff just copies all the relevant components to the top level dir.

We'll do this as a small migration, which first copies, then later deletes the old versions. This lets us iterate slightly on the components to get them working in a standalone fashion.

Reviewed By: zzl0

Differential Revision: D59305300

fbshipit-source-id: 55fddfb0188ea96d196816a51e8693fc8f49ae38