changesets: remove generation numbers from argument to add_many
Summary:
Currently `Changesets::add_many` requires providing the generation number for all inserted changesets. However, the generation numbers can be calculated in the implementation of `add_many` by querying the generation number of all already inserted parents (which is usually one parent) which doesn't add much overhead and is what we currently do in `Changesets::add` which we use when saving any new changeset in Mononoke.

This diff removes the generation number arguments and adds the logic to calculate them in `Changesets::add_many`.

Reviewed By: mitrandir77

Differential Revision: D59584635

fbshipit-source-id: f6c3e9479eda0a13519e89c27b4655111afcfae7