Bandwidth instrumentation(2/2): Add a new event SaplingBlobDownloadEvent in LogEvent.h
Summary:
# Background:
Verbatim feedback is that the Eden does not work well with low bandwidth scenarios. Hypothesis is that users would run `eden doctor` more frequently when bandwidth is low.

# Solution:
To validate this hypothesis and determine next steps to address low bandwidth scenarios, instrumenting file size along with download duration for each downloaded file to get a closer signal to bandwidth.

# Implementation
1. Add a new event `SaplingBlobDownloadEvent` in LogEvent.h**
2. **Use `SaplingBlobDownloadEvent` event to instrument duration and size in scuba**

# Tracking:
[Scuba](https://fburl.com/scuba/edenfs_events/d1lfpufx)

Differential Revision: D59301287

fbshipit-source-id: 980aade9510f6ca0baba3f94ac0a9208d85775d1