goto: flush stores better after calling into Rust goto
Summary: Previously if Python called into Rust "goto" and did not trigger the Python repo.manifestlog or repo.fileslog objects to be loaded, repo.commitpending was not flushing the files or trees written by the Rust goto. Fix by flushing the Rust repo object if either of the Python stores were not loaded.

Reviewed By: zzl0

Differential Revision: D59588261

fbshipit-source-id: e775528a55c054f8253d78fe769346341f7aaafe